+++ packages/svelte/src/lib/components/toast/toast.stories.ts
@@ -1,6 +1,8 @@
 import type { Meta } from '@storybook/svelte'
 import ActionExample from './examples/action.svelte'
 import BasicExample from './examples/basic.svelte'
+import PromiseExample from './examples/promise-toast.svelte'
+import TypesExample from './examples/types.svelte'
 import UpdateExample from './examples/update.svelte'
 
 const meta: Meta = {
@@ -18,6 +20,18 @@ export const Basic = {
 export const Action = {
   render: () => ({
     Component: ActionExample,
+  }),
+}
+
+export const PromiseToast = {
+  render: () => ({
+    Component: PromiseExample,
+  }),
+}
+
+export const Types = {
+  render: () => ({
+    Component: TypesExample,
   }),
 }