+++ packages/vue/src/components/toast/toast.stories.vue
@@ -1,6 +1,8 @@
 <script setup lang="ts">
 import Action from './examples/action.vue'
 import Basic from './examples/basic.vue'
+import PromiseToast from './examples/promise-toast.vue'
+import Types from './examples/types.vue'
 import Update from './examples/update.vue'
 </script>
 
@@ -8,6 +10,12 @@ import Update from './examples/update.vue'
   <Story title="Toast">
     <Variant title="Basic">
       <Basic />
+    </Variant>
+    <Variant title="PromiseToast">
+      <PromiseToast />
+    </Variant>
+    <Variant title="Types">
+      <Types />
     </Variant>
     <Variant title="Action">
       <Action />